style: dependency key 위치 변경
